backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
replicate==0.11.0
mangum==0.17.0
pydantic==2.4.2 